+++ docs/scripts/controllers/main.controller.js
@@ -15,21 +15,19 @@
 		.controller('mainController', mainController);
 
 	mainController.$inject = [
-		'$scope',
 		'logService',
 		'dateService'
 	];
 
-	function mainController($scope, logService, dateService) {
+	function mainController(logService, dateService) {
 		const vm = this;
 
 		// Public data
 		vm.data = {
-			controller      : 'mainController',
-			today           : new Date(),
-			initialDate     : null,
-			initialDateError: false,
-			showDatepicker  : false
+			controller    : 'mainController',
+			today         : new Date(),
+			initialDate   : null,
+			showDatepicker: false
 		};
 
 		// Public methods
@@ -42,25 +40,6 @@
 		// Init
 		logService.init(vm.data.controller);
 		vm.calculatedDate = null;
-
-		// Watch the initial date change
-		// To display UI message when an error occurred
-		$scope.$watch('initialDate', $value => {
-			logService.log($value);
-			try {
-				vm.data.initialDate = new Date($value);
-			}
-			catch ($e) {
-			}
-			logService.log(vm.data.initialDate);
-
-			if (!vm.data.initialDate) {
-				vm.data.initialDateError = 'This is not a valid date';
-			}
-			else {
-				vm.data.initialDateError = false;
-			}
-		});
 
 		function define21Date() {
 			logService.fnCalled('define21Date');